+++ packages/nestjs/jest.config.js
@@ -1,4 +1,4 @@
 module.exports = {
-  preset: '@nova-platform/jest-preset',
+  preset: '@ahincho/nova-jest-preset',
   setupFiles: ['reflect-metadata'],
 };